MERLIN for causal heterogeneity using summary statistics

MERLIN (Causal Heterogeneity using Summary Statistics) is a unified Bayesian framework that jointly estimates average and heterogeneity effects using summary data from genome-wide association studies (GWAS) and Genome-Wide Interaction Studies (GWIS).

Details

The DESCRIPTION file: This package was not yet installed at build time.
Index: This package was not yet installed at build time.

This section should provide a more detailed overview of how to use the package, including the most important functions.

Author

First Last [aut, cre] (YOUR-ORCID-ID)

Maintainer: First Last <first.last@example.com>

References

This optional section can contain literature or other references for background information.

See also

Examples

  ## Optional simple examples of the most important functions
  ## Use \dontrun{} around code to be shown but not executed